@@ -23,6 +23,7 @@
 
 const std::string serviceId = "service-sample";
 const std::string clientId = "client-sample";
+const std::string otherclientId = "other-client-sample";
 
 const std::string domain = "local";
 const std::string testAddress = "commonapi.communication.TestInterface";
@@ -67,6 +68,7 @@ protected:
         ASSERT_TRUE(serviceRegistered);
 
         testProxy_ = runtime_->buildProxy<TestInterfaceProxy>(domain, testAddress, clientId);
+        ASSERT_TRUE((bool)testProxy_);
         int i = 0;
         while(!testProxy_->isAvailable() && i++ < 100) {
             std::this_thread::sleep_for(std::chrono::milliseconds(10));
@@ -81,14 +83,16 @@ protected:
 
         ASSERT_TRUE(serviceUnregistered);
 
-        // wait that proxy is not available
-        int counter = 0;  // counter for avoiding endless loop
-        do {
-            std::this_thread::sleep_for(std::chrono::microseconds(tasync*wf));
-            counter++;
-        } while ( testProxy_->isAvailable() && counter < 100 );
+        if (testProxy_) {
+            // wait that proxy is not available
+            int counter = 0;  // counter for avoiding endless loop
+            do {
+                std::this_thread::sleep_for(std::chrono::microseconds(tasync*wf));
+                counter++;
+            } while ( testProxy_->isAvailable() && counter < 100 );
 
-        ASSERT_FALSE(testProxy_->isAvailable());
+            ASSERT_FALSE(testProxy_->isAvailable());
+        }
     }
 
     uint8_t value_;
@@ -105,6 +109,7 @@ TEST_F(CMBroadcasts, NormalBroadcast) {
 
     CommonAPI::CallStatus callStatus;
     std::atomic<uint8_t> result;
+    std::atomic<CommonAPI::CallStatus> subStatus;
     result = 0;
 
     // subscribe to broadcast
@@ -112,7 +117,19 @@ TEST_F(CMBroadcasts, NormalBroadcast) {
         const uint8_t &y
     ) {
         result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus = status;
     });
+
+    // check that subscription has succeeded
+    for (int i = 0; i < 100; i++) {
+        if (subStatus == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus);
 
     // send value '1' via a method call - this tells stub to broadcast
     uint8_t in_ = 1;
@@ -254,6 +271,7 @@ TEST_F(CMBroadcasts, SelectiveBroadcast) {
 TEST_F(CMBroadcasts, BroadcastStubGoesOfflineOnlineAgain) {
     CommonAPI::CallStatus callStatus;
     std::atomic<uint8_t> result;
+    std::atomic<CommonAPI::CallStatus> subStatus;
     result = 0;
 
     runtime_->unregisterService(domain, CMBroadcastsStub::StubInterface::getInterface(),
@@ -272,18 +290,30 @@ TEST_F(CMBroadcasts, BroadcastStubGoesOfflineOnlineAgain) {
         const uint8_t &y
     ) {
         result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus = status;
     });
 
     bool serviceRegistered = runtime_->registerService(domain, testAddress, testStub_, serviceId);
     ASSERT_TRUE(serviceRegistered);
 
-    // wait that proxy is  available
+    // wait that proxy is available
     counter = 0;  // counter for avoiding endless loop
     while ( !testProxy_->isAvailable() && counter < 100 ) {
         std::this_thread::sleep_for(std::chrono::microseconds(tasync));
         counter++;
     }
     ASSERT_TRUE(testProxy_->isAvailable());
+
+    // check that subscription has succeeded
+    for (int i = 0; i < 100; i++) {
+        if (subStatus == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus);
 
     // send value '1' via a method call - this tells stub to broadcast
     uint8_t in_ = 1;
@@ -297,6 +327,8 @@ TEST_F(CMBroadcasts, BroadcastStubGoesOfflineOnlineAgain) {
         std::this_thread::sleep_for(std::chrono::microseconds(tasync));
     }
     EXPECT_EQ(result, 1);
+
+    subStatus = CommonAPI::CallStatus::UNKNOWN;
 
     runtime_->unregisterService(domain, CMBroadcastsStub::StubInterface::getInterface(),
                             testAddress);
@@ -320,6 +352,13 @@ TEST_F(CMBroadcasts, BroadcastStubGoesOfflineOnlineAgain) {
         counter++;
     }
     ASSERT_TRUE(testProxy_->isAvailable());
+
+    // check that subscription has succeeded
+    for (int i = 0; i < 100; i++) {
+        if (subStatus == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus);
 
     result = 0;
     // send value '1' via a method call - this tells stub to broadcast
@@ -395,6 +434,9 @@ TEST_F(CMBroadcasts, SelectiveBroadcastStubGoesOfflineOnlineAgain) {
     }
     ASSERT_TRUE(testProxy_->isAvailable());
 
+    // Let the proxy re-subscribe before trigger a notify one
+    std::this_thread::sleep_for(std::chrono::microseconds(tasync * 2));
+
     // send value '3' via a method call - this tells stub to broadcast selective
     in_ = 3;
     out_ = 0;
@@ -435,6 +477,9 @@ TEST_F(CMBroadcasts, SelectiveBroadcastStubGoesOfflineOnlineAgain) {
     testProxy_->testMethod(in_, callStatus, out_);
     EXPECT_EQ(callStatus, CommonAPI::CallStatus::SUCCESS);
 
+    // Let the proxy re-subscribe before trigger a notify
+    std::this_thread::sleep_for(std::chrono::microseconds(tasync * 2));
+
     result = 0;
     // send value '3' via a method call - this tells stub to broadcast selective
     in_ = 3;
@@ -454,6 +499,7 @@ TEST_F(CMBroadcasts, SelectiveBroadcastStubGoesOfflineOnlineAgain) {
 TEST_F(CMBroadcasts, NormalBroadcast_Two_proxies_subscribe_and_one_reset) {
     CommonAPI::CallStatus callStatus;
     std::atomic<uint8_t> result, result2;
+    std::atomic<CommonAPI::CallStatus> subStatus, subStatus2;
     result = 0;
     result2 = 0;
 
@@ -469,14 +515,32 @@ TEST_F(CMBroadcasts, NormalBroadcast_Two_proxies_subscribe_and_one_reset) {
         const uint8_t &y
     ) {
         result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus = status;
     });
-
     // subscribe to broadcast
     anotherProxy->getBTestEvent().subscribe([&](
         const uint8_t &y
     ) {
         result2 = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus2 = status;
     });
+
+    // check that all subscriptions have succeeded
+    for (int i = 0; i < 100; i++) {
+        if (subStatus == CommonAPI::CallStatus::SUCCESS &&
+                subStatus2 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus);
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus2);
 
     // send value '1' via a method call - this tells stub to broadcast
     uint8_t in_ = 1;
@@ -522,11 +586,19 @@ TEST_F(CMBroadcasts, NormalBroadcast_Two_proxies_subscribe_and_one_reset) {
     }
     ASSERT_TRUE(anotherProxy->isAvailable());
 
+    subStatus = CommonAPI::CallStatus::UNKNOWN;
+    subStatus2 = CommonAPI::CallStatus::UNKNOWN;
+
     // subscribe to broadcast
     testProxy_->getBTestEvent().subscribe([&](
         const uint8_t &y
     ) {
         result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus = status;
     });
 
     // subscribe to broadcast
@@ -534,7 +606,21 @@ TEST_F(CMBroadcasts, NormalBroadcast_Two_proxies_subscribe_and_one_reset) {
         const uint8_t &y
     ) {
         result2 = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus2 = status;
     });
+
+    // check that all subscriptions have succeeded
+    for (int i = 0; i < 100; i++) {
+        if (subStatus == CommonAPI::CallStatus::SUCCESS &&
+                subStatus2 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus);
+    EXPECT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus2);
 
     // send value '1' via a method call - this tells stub to broadcast
     in_ = 1;
@@ -549,6 +635,141 @@ TEST_F(CMBroadcasts, NormalBroadcast_Two_proxies_subscribe_and_one_reset) {
     }
     EXPECT_EQ(result, 1);
     EXPECT_EQ(result2, 1);
+}
+
+TEST_F(CMBroadcasts, Two_proxies_subscribe_delete_one_proxy_status_listener_test) {
+    std::atomic<CommonAPI::CallStatus> subStatus1;
+    std::atomic<CommonAPI::CallStatus> subStatus2;
+    std::atomic<CommonAPI::CallStatus> subStatus3;
+    std::atomic<CommonAPI::CallStatus> subStatus4;
+    std::atomic<uint8_t> result, result2;
+    result = 0;
+    result2 = 0;
+
+    // subscribe first Proxy
+    subStatus1 = CommonAPI::CallStatus::UNKNOWN;
+    testProxy_->getBTestEvent().subscribe([&](
+        const uint8_t &y
+    ) {
+        result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus1 = status;
+    });
+    // check that status was correctly received
+    for (int i = 0; i < 100; i++) {
+        if (subStatus1 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus1);
+
+    // subscribe second Proxy (another connection)
+    subStatus2 = CommonAPI::CallStatus::UNKNOWN;
+    auto testProxy2_ = runtime_->buildProxy<TestInterfaceProxy>(domain, testAddress, clientId);
+    testProxy2_->getBTestEvent().subscribe([&](
+        const uint8_t &y
+    ) {
+        result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus2 = status;
+    });
+    // check that status was correctly received
+    for (int i = 0; i < 100; i++) {
+        if (subStatus2 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus2);
+
+    // Build second proxy on same connection than first proxy
+    auto secondProxy = runtime_->buildProxy<TestInterfaceProxy>(domain, testAddress, clientId);
+    int i = 0;
+    while(!secondProxy->isAvailable() && i++ < 100) {
+        std::this_thread::sleep_for(std::chrono::milliseconds(10));
+    }
+    ASSERT_TRUE(secondProxy->isAvailable());
+
+    // subscribe second Proxy same connection
+    subStatus3 = CommonAPI::CallStatus::UNKNOWN;
+    secondProxy->getBTestEvent().subscribe([&](
+        const uint8_t &y
+    ) {
+        result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus3 = status;
+    });
+
+    // Build second proxy on same connection than first proxy
+    auto thirdProxy = runtime_->buildProxy<TestInterfaceProxy>(domain, testAddress, otherclientId);
+    i = 0;
+    while(!thirdProxy->isAvailable() && i++ < 100) {
+        std::this_thread::sleep_for(std::chrono::milliseconds(10));
+    }
+    ASSERT_TRUE(thirdProxy->isAvailable());
+
+    // subscribe second Proxy same connection
+    subStatus4 = CommonAPI::CallStatus::UNKNOWN;
+    thirdProxy->getBTestEvent().subscribe([&](
+        const uint8_t &y
+    ) {
+        result = y;
+    },
+    [&](
+        const CommonAPI::CallStatus &status
+    ) {
+        subStatus4 = status;
+    });
+
+    // Delete first proxy to ensure same and other connection receive their status anyways!
+    testProxy_.reset();
+
+    // check that status was correctly received anyways
+    for (int i = 0; i < 100; i++) {
+        if (subStatus3 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus3);
+
+    // check that status was correctly received anyways
+    for (int i = 0; i < 100; i++) {
+        if (subStatus4 == CommonAPI::CallStatus::SUCCESS) break;
+        std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+    }
+    ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus4);
+
+    subStatus1 = CommonAPI::CallStatus::UNKNOWN;
+    subStatus2 = CommonAPI::CallStatus::UNKNOWN;
+    subStatus3 = CommonAPI::CallStatus::UNKNOWN;
+    subStatus4 = CommonAPI::CallStatus::UNKNOWN;
+
+    bool serviceUnregistered =
+            runtime_->unregisterService(domain, CMBroadcastsStub::StubInterface::getInterface(),
+                    testAddress);
+     ASSERT_TRUE(serviceUnregistered);
+
+     std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+
+     bool serviceRegistered = runtime_->registerService(domain, testAddress, testStub_, serviceId);
+     ASSERT_TRUE(serviceRegistered);
+
+     // check that all states were correctly received after service comes up again
+     for (int i = 0; i < 100; i++) {
+         if (subStatus2 == CommonAPI::CallStatus::SUCCESS &&
+                 subStatus4 == CommonAPI::CallStatus::SUCCESS &&
+                 subStatus4 == CommonAPI::CallStatus::SUCCESS) break;
+         std::this_thread::sleep_for(std::chrono::microseconds(tasync));
+     }
+     ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus2);
+     ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus3);
+     ASSERT_EQ(CommonAPI::CallStatus::SUCCESS, subStatus4);
+     ASSERT_EQ(CommonAPI::CallStatus::UNKNOWN, subStatus1);
 }
 
 int main(int argc, char** argv) {
